Add hashed branch name to the generated version
Needed something to disambiguate branches named "dev", "my-dev", etc.
Considered a delimiter of some sort, but with such a delimiter there
is always the risk of that delimiter being used in the branch name
itself (and bugs as a result). This approach provides a delimiter
that has a low risk of being used in other branch names, while also
being deterministic.

Note that I kept the branch name in the version (as opposed to including
only the hash alone) for human-readability and being able to track
versions back to a particular branch.

Also note that the first-7-of-sha1 is how git determines its shortened
hashes; there is enough variability within the scope of a single repo
that collisions are low-to-non-existent. Since the version is being
generated using git, this approach (i.e. git-7) felt right.