Option in upload command to write the version back to local config
This is helpful for repos that control an api document (and are thus
uploading it to Apibuilder) but also use the same document to generate
code locally in the repo. Using "latest" as the version doesn't work
for non-semver (i.e. "development") versions uploaded to Apibuilder.
Also, doing something like "most recent uploaded version" does not
work for repos with multiple people working on multiple branches. So
it is easiest to store the version in the local config and automatically
update it as part of the workflow.